Различные валидации и оптимизации после тестирования.

+++ backend/foodgram_dj/userprofile/views.py
@@ -3,13 +3,10 @@ from django.shortcuts import get_object_or_404
 from rest_framework.decorators import action
 from rest_framework.response import Response
 from rest_framework.validators import ValidationError
-from rest_framework import viewsets, pagination, mixins, permissions, status
+from rest_framework import pagination, permissions, status
 from djoser import views
 from .models import UserProfile, Subscription
-from .serializers import UserProfileSerializer  #, SubscriptionSerializer
-
-
-
+from .serializers import UserProfileSerializer
 
 
 class UserProfileViewSet(views.UserViewSet):
@@ -18,22 +15,22 @@ class UserProfileViewSet(views.UserViewSet):
     pagination_class = pagination.LimitOffsetPagination
     permission_classes = [permissions.IsAuthenticatedOrReadOnly, ]
 
-    @action(detail=False, methods=['get'], url_path='me')
+    @action(detail=False,
+            methods=['get'],
+            url_path='me',
+            permission_classes=[permissions.IsAuthenticated])
     def get_current_user(self, request):
         """Метод для получения текущего пользователя"""
         return Response(self.get_serializer(request.user).data)
 
     @action(detail=True, methods=['get'])
     def get_user(self, request, pk=None):
-        # user = get_object_or_404(UserProfile, pk=pk)
         user = self.get_object()
-        # is_subscribed = Subscription.objects.filter(
-        #     follows=user.pk,
-        #     user=request.user.pk
-        # ).exists()
+        if not user:
+            return Response({'detail': 'Страница не найдена.'},
+                            status=status.HTTP_404_NOT_FOUND)
         serializer = self.get_serializer(user)
         data = serializer.data
-        # data['is_subscribed'] = is_subscribed
         return Response(data, status=status.HTTP_200_OK)
 
     @action(detail=False, methods=['put', 'delete'], url_path='me/avatar')
@@ -45,6 +42,12 @@ class UserProfileViewSet(views.UserViewSet):
                                              partial=True)
             serializer.is_valid(raise_exception=True)
 
+            if 'avatar' not in request.data or not request.data.get('avatar'):
+                return Response(
+                    {'detail': 'Поле \'avatar\' не задано!'},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+
             if user.avatar:
                 user.avatar.delete()
 
@@ -52,6 +55,11 @@ class UserProfileViewSet(views.UserViewSet):
             return Response(
                 {'avatar': serializer.data['avatar']},
                 status=status.HTTP_200_OK
+            )
+        if not user.avatar:
+            return Response(
+                {'detail': 'Аватар не существует!'},
+                status=status.HTTP_400_BAD_REQUEST
             )
         user.avatar.delete()
         user.save()
@@ -81,7 +89,16 @@ class UserProfileViewSet(views.UserViewSet):
                 status=status.HTTP_201_CREATED
             )
 
-        get_object_or_404(Subscription, user=request.user, follows=to_sub
-                          ).delete()
+        # count, var = get_object_or_404(Subscription, user=request.user,
+        #                                follows=to_sub
+        #                                ).delete()
+        count, var = Subscription.objects.filter(user=request.user,
+                                                 follows=to_sub
+                                                 ).delete()
+        if count == 0:
+            return Response(
+                {'detail': 'Ошибка отписки: не был подписан!'},
+                status=status.HTTP_400_BAD_REQUEST
+            )
 
         return Response(status=status.HTTP_204_NO_CONTENT)
